[S07P12A506-170][BE] feature: add Likepost

See merge request s07-webmobile2-sub2/S07P12A506!79

+++ back/Trippy/src/main/java/com/ssafy/trippy/Service/Impl/BookmarkServiceImpl.java
@@ -36,6 +36,9 @@ public class BookmarkServiceImpl implements BookmarkService {
 
     @Override
     public void deleteBookmark(RequestBookmarkDto requestBookmarkDto) {
+        if(!bookmarkRepository.existsByMemberIdAndCommunityPostId(requestBookmarkDto.getMemberId(),requestBookmarkDto.getCommunityPostId())){
+            throw new IllegalArgumentException("존재하지 않는 북마크");
+        }
         bookmarkRepository.deleteByMemberIdAndCommunityPostId(requestBookmarkDto.getMemberId(),requestBookmarkDto.getCommunityPostId());
     }
 }
